# cabinet-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CabinetPro API",
        "description": "API REST de gestion de cabinet d'avocat — CabinetPro.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/evenements": {
            "get": {
                "tags": [
                    "Calendrier"
                ],
                "summary": "Liste des évènements",
                "operationId": "67e6473ed23851b615eb46e892c828e5",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagination des évènements"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Calendrier"
                ],
                "summary": "Créer un évènement",
                "operationId": "75bea6238c4386b83a5d9ce897bbebf5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Evenement"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Évènement créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/evenements/{evenement}": {
            "get": {
                "tags": [
                    "Calendrier"
                ],
                "summary": "Détail d'un évènement",
                "operationId": "a0c8249285e62bac2e3343cac8c65952",
                "parameters": [
                    {
                        "name": "evenement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Évènement"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Calendrier"
                ],
                "summary": "Modifier un évènement",
                "operationId": "73e214a5884dc0d4abd85112bdaa8a1a",
                "parameters": [
                    {
                        "name": "evenement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Evenement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Évènement modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Calendrier"
                ],
                "summary": "Supprimer un évènement",
                "operationId": "b1c038a2ca32cd3a58d6338c6a2b99b8",
                "parameters": [
                    {
                        "name": "evenement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Évènement supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Evenement": {
                "required": [
                    "titre",
                    "type",
                    "debut"
                ],
                "properties": {
                    "titre": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "rendez_vous",
                            "audience",
                            "echeance"
                        ]
                    },
                    "debut": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fin": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "dossier_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "client_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "lieu": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}